refacto(front/viewer): split logic in many component
