+++ src/frontend/classic/components/ThumbnailViewerImage.svelte
