+++ src/frontend/classic/components/Viewer.svelte
@@ -1,15 +1,8 @@
 <script lang="ts">
-    import ChevronLeft24 from "carbon-icons-svelte/lib/ChevronLeft24";
-    import ChevronRight24 from "carbon-icons-svelte/lib/ChevronRight24";
-    import ZoomIn24 from "carbon-icons-svelte/lib/ZoomIn24";
-    import ZoomOut24 from "carbon-icons-svelte/lib/ZoomOut24";
-    import RowInsert24 from "carbon-icons-svelte/lib/RowExpand24";
-    import RowDelete24 from "carbon-icons-svelte/lib/RowCollapse24";
-    import Misuse24 from "carbon-icons-svelte/lib/Misuse24";
-    import { createEventDispatcher } from "svelte";
-    import ViewerThumbnailImage from "./ViewerThumbnailImage.svelte";
-    import ViewerImageWide from "./ViewerImageWide.svelte";
     import type { IMediaContainer } from "../../../engine/providers/MediaPlugin";
+    import ThumbnailViewer from "./ThumbnailViewer.svelte";
+    import WideViewer from "./WIdeViewer.svelte";
+    import VideoViewer from "./VideoViewer.svelte";
 
     type Mode = "Thumbnail" | "Wide" | "Video";
 
@@ -17,130 +10,35 @@
     export let mode: Mode = "Thumbnail";
     let throttlingDelay = 1000;
 
-    const dispatch = createEventDispatcher();
-
     let update: Promise<void> | undefined;
     $: update = item?.Update();
 
-    let imageWidth = 75;
-    let imagePadding = 2;
-
-    function increaseImagePadding() {}
-    function decreaseImagePadding() {}
-    function zoomIn() {}
-    function zoomOut() {}
     function toggleThumbnailViewer() {
         mode = "Thumbnail";
-        const rootdiv: any = document.getElementById("Content");
-        const contentdiv: any = document.getElementById("Viewer");
-        rootdiv.appendChild(contentdiv);
     }
     function toggleWideViewer() {
         mode = "Wide";
-        const contentdiv: any = document.getElementById("Viewer");
-        document.body.appendChild(contentdiv);
-    }
-    function toggleFullScreen() {
-        if (!document.fullscreenElement) {
-            document.documentElement.requestFullscreen();
-        } else if (document.exitFullscreen) {
-            document.exitFullscreen();
-        }
     }
 </script>
 
 <div id="Viewer">
-    {#if mode === "Thumbnail" || mode === "Wide"}
-        <div id="Contents" class={mode}>
-            {#await update}
-                <p>...loading items</p>
-            {:then}
-                {#if mode === "Wide"}
-                    <div
-                        id="Buttons"
-                        tabindex="0"
-                        on-blur="focus()"
-                        on-keydown="onKeyDown"
-                    >
-                        <span class="title"
-                            >{item?.Parent?.Title ?? "unkown"}</span
-                        >
-                        <ChevronLeft24
-                            label="Item Down (ArrowLeft)"
-                            on:click={(e) => dispatch("requestItemDown", item)}
-                        />
-                        <ChevronRight24
-                            label="Item Up (ArrowRight)"
-                            on:click={(e) => dispatch("requestItemUp", item)}
-                        />
-                        &nbsp;
-                        <RowDelete24
-                            label="Decrease spacing between images (CTRL ➖)"
-                            on:click={decreaseImagePadding}
-                        />
-                        <RowInsert24
-                            label="Increase spacing between images (CTRL ➕)"
-                            on:click={increaseImagePadding}
-                        />
-                        &nbsp;
-                        <ZoomIn24 label="Zoom In (➕)" on:click={zoomIn} />
-                        <ZoomOut24 label="Zoom Out (➖)" on:click={zoomOut} />
-                        &nbsp
-                        <Misuse24
-                            label="Close (ESC)"
-                            on:click={toggleThumbnailViewer}
-                        />
-                    </div>
-                {/if}
-                {#each item.Entries as content, index}
-                    <!--{#await contentfetch.GetImage()}
-                        <p>...loading image</p>
-                    {:then content}-->
-                    {#if mode === "Thumbnail"}
-                        <ViewerThumbnailImage
-                            src={content.SourceURL}
-                            throttlingDelay={throttlingDelay * index + 1}
-                            handleClick={toggleWideViewer}
-                            title="Page {index}"
-                        />
-                    {:else if mode === "Wide"}
-                        <ViewerImageWide
-                            alt="content_{index}"
-                            src={content.SourceURL}
-                            {imagePadding}
-                            {imageWidth}
-                            throttlingDelay={throttlingDelay * index}
-                        />
-                    {/if}
-                    <!--
-                    {:catch error}
-                        <p class="error">{error.message}</p>
-                    {/await}-->
-                {/each}
-            {:catch error}
-                <p class="error">Unable to load item : {error}</p>
-            {/await}
-        </div>
-    {:else if mode === "Video"}
-        <!-- WIP nothing to see here -->
-        <div id="FullScreen" on-dblclick="toggleFullscreen">
-            <!-- disablePictureInPicture controlsList="nodownload nofullscreen" -->
-            <video id="Video" controls>
-                <!-- dummy tracks to show subtitle menu in video player -->
-                <template is="dom-repeat" items="[[ media.subtitles ]]">
-                    <!-- dummy tracks to show subtitle menu in video player -->
-                    <track
-                        kind="captions"
-                        src="data:text/vtt,WEBVTT"
-                        label="[[ item.locale ]]"
-                        srclang="[[ item.locale ]]"
-                    />
-                </template>
-            </video>
-        </div>
-    {:else}
-        Unknown mode requested
-    {/if}
+    <div id="Contents" class={mode}>
+        {#await update}
+            <p>...loading items</p>
+        {:then}
+            {#if mode === "Thumbnail"}
+                <ThumbnailViewer {item} {throttlingDelay} {toggleWideViewer} />
+            {:else if mode === "Wide"}
+                <WideViewer {item} {throttlingDelay} {toggleThumbnailViewer} />
+            {:else if mode === "Video"}
+                <VideoViewer />
+            {:else}
+                Unknown mode requested
+            {/if}
+        {:catch error}
+            <p class="error">Unable to load item : {error}</p>
+        {/await}
+    </div>
 </div>
 
 <style>
@@ -214,20 +112,6 @@
     :global(#Viewer .button) {
         cursor: pointer;
         margin: 0.25em;
-    }
-    #FullScreen {
-        width: 100%;
-        height: 100%;
-        background-color: var(--cds-ui-01);
-    }
-    #Video {
-        width: 100%;
-        height: 100%;
-        object-fit: contain;
-        outline: none !important;
-    }
-    video::-webkit-media-controls-fullscreen-button {
-        display: none;
     }
 
     /* hack to allow video to expand size when wrapped in subtitle container
