+++ src/frontend/classic/components/WideViewerImage.svelte
